refactor: remove app.ts
