--- app/app.ts
@@ -1,5 +1,0 @@
-import React from "react"
-
-const AppContext = React.createContext<any>(null)
-
-export default AppContext
